backend/requirements.txt
# requirements.txt
streamlit==1.24.0
langchain-openai>=0.2.6
langchain-community>=0.0.20
langchain-core>=0.3.15
openai>=1.0.0
python-dotenv==1.0.0
pypdf2==3.0.1
azure-storage-blob==12.16.0
chromadb==0.4.15
pandas==2.0.0 
openpyxl==3.1.2
tiktoken>=0.8.0
unstructured
azure-search-documents>=11.4.0
azure-identity>=1.12.0
rank-bm25>=0.2.2
scikit-learn>=1.3.0
pysqlite3-binary